feat: new tests

+++ src/handlers/users/deleteUser.ts
@@ -1,5 +1,6 @@
 import type { Schema } from '@march1-org/db-template';
 import { eq } from 'drizzle-orm';
+import { error } from 'elysia';
 import type Redis from 'ioredis';
 import type { DbType } from 'lib/db';
 
@@ -16,6 +17,15 @@ export async function deleteUser({
   params: { id },
   cache,
 }: Options) {
+  const user = await db
+    .select()
+    .from(users)
+    .where(eq(users.id, Number(id)));
+
+  if (user.length === 0) {
+    return error('Not Found');
+  }
+
   await db.delete(users).where(eq(users.id, Number(id)));
 
   const cacheKey = `user:${id}`;
